requirements.txt
Django==1.5.8
Fabric==1.3.0
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
PyYAML==3.11
SQLAlchemy==0.9.6
amqp==1.4.5
anyjson==0.3.3
apache-libcloud==0.15.1
billiard==3.3.0.18
boto==2.9.2
ceiclient==0.2.2
celery==3.1.12
cloudinitd==1.3.1
dashi==0.3.0
django-easysettings==1.0.1
django-password-policies==0.2
django-statsd==1.9.1
django-tokenapi==0.2.0
kombu==3.0.21
pycrypto==2.3
python-statsd==1.6.3
pytz==2014.4
simplejson==3.5.3
ssh==1.8.0
uuid==1.30
